Move AppContext to top level
