+++ frontend/src/App.tsx
@@ -1,9 +1,8 @@
-import React, { Suspense, useState, useEffect } from 'react';
+import React, { Suspense } from 'react';
 import ReactDOM from 'react-dom/client';
 import Main from './basic/Main';
 import { CssBaseline } from '@mui/material';
 import { AppContextProvider } from './contexts/AppContext';
-import { ThemeProvider, createTheme, Theme } from '@mui/material/styles';
 import UnsafeAlert from './components/UnsafeAlert';
 import TorConnectionBadge from './components/TorConnection';
 
@@ -11,46 +10,18 @@ import { I18nextProvider } from 'react-i18next';
 import i18n from './i18n/Web';
 
 import { systemClient } from './services/System';
-import { Settings } from './models';
 import ErrorBoundary from './components/ErrorBoundary';
 
-const makeTheme = function (settings: Settings) {
-  const theme: Theme = createTheme({
-    palette: {
-      mode: settings.mode,
-      background: {
-        default: settings.mode === 'dark' ? '#070707' : '#fff',
-      },
-    },
-    typography: { fontSize: settings.fontSize },
-  });
-
-  return theme;
-};
-
 const App = (): JSX.Element => {
-  const [theme, setTheme] = useState<Theme>(makeTheme(new Settings()));
-  const [settings, setSettings] = useState<Settings>(new Settings());
-
-  useEffect(() => {
-    setTheme(makeTheme(settings));
-  }, [settings.fontSize, settings.mode]);
-
-  useEffect(() => {
-    i18n.changeLanguage(settings.language);
-  }, []);
-
   return (
     <ErrorBoundary>
-      <Suspense fallback='loading language'>
+      <Suspense fallback='loading'>
         <I18nextProvider i18n={i18n}>
-          <ThemeProvider theme={theme}>
-            <AppContextProvider settings={settings} setSettings={setSettings}>
-              <CssBaseline />
-              {window.NativeRobosats === undefined ? <UnsafeAlert /> : <TorConnectionBadge />}
-              <Main />
-            </AppContextProvider>
-          </ThemeProvider>
+          <AppContextProvider>
+            <CssBaseline />
+            {window.NativeRobosats === undefined ? <UnsafeAlert /> : <TorConnectionBadge />}
+            <Main />
+          </AppContextProvider>
         </I18nextProvider>
       </Suspense>
     </ErrorBoundary>
@@ -58,6 +29,7 @@ const App = (): JSX.Element => {
 };
 
 const loadApp = () => {
+  // waits until the environment is ready for the Android WebView app
   if (systemClient.loading) {
     setTimeout(loadApp, 200);
   } else {
